cleanup: isolate Groove runtime state
